requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
rq==1.15.1
python-jose[cryptography]==3.3.0
huggingface-hub==0.19.4
pydantic==2.5.0